kronos-stock-predict/prediction/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
pyyaml>=6.0
requests>=2.31.0
